[torch] Use __prepare_scriptable__ for closures
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120806

This fixes a case left incomplete by https://github.com/pytorch/pytorch/pull/106229
The object is using __prepare_scriptable__ correctly inside of torch.jit.script()
but the clousre that is obtained below is using the non-prepared version.
This causes issues when the prepared and non-prepared versions are in different python modules.

The next diff fails without this fix.
ghstack-source-id: 217307158
@exported-using-ghexport

Differential Revision: [D54308741](https://our.internmc.facebook.com/intern/diff/D54308741/)