[dtensor] to_local backward grad placement passthrough
to_local accepts a `grad_placements` if user choose to pass, previously
we enforce the grad_out to be the "same" placement as the current
DTensor for safety.

But I realized that we DO NOT need to enforce this constraint. Why?
backward placement does not need to be the same as fwd tensor placement, this
is already the case for param vs param.grad (i.e. param can be replicate
and grad can be partial), so we should not restrict this to activation
vs activation grad too

ghstack-source-id: bdc63440b27a9d196172239aeadd0edd46253b14
Pull Request resolved: https://github.com/pytorch/pytorch/pull/121474